Shorten section on planning for mistakes
This section was far longer than any of the other sections and got
into a lot more detail. In addition, the editors specifically asked
us to remove the second level of headings, which wasn't tenable
without shortening the content.

Three subsections were removed entirely:

1. Use a variety of oracles. The spirit of this was encorporated
into the current language, but not the labeled practice itself.

2. Refactoring legacy code for testing. This should be moved to
the new 'What's next' section and the existing language has been
moved there for editing.

3. TDD. If we are going to keep this (and it is pretty equivocal at
the moment) we should move it to the new 'What's next' section and
the existing language has been moved there for editing.